+++ src/main/java/de/tum/cit/aet/artemis/core/config/builder/properties/LiquibaseEnabledProperty.java
@@ -1,6 +1,6 @@
 package de.tum.cit.aet.artemis.core.config.builder.properties;
 
-import static de.tum.cit.aet.artemis.core.config.conditions.ConditionHelper.isBuildAgentOnlyMode;
+import static de.tum.cit.aet.artemis.core.config.conditions.ArtemisConfigHelper.isBuildAgentOnlyMode;
 
 import org.springframework.core.env.ConfigurableEnvironment;
